requirements.txt
Flask==2.3.2
pymongo==4.5.1
bson==0.5.10
gunicorn==20.1.0
python-dotenv==1.1.0


